requirements.txt
openai
anthropic
google-genai
rich
prompt_toolkit
requests
pyttsx3
duckduckgo_search
trafilatura
